We are creating a discord-like application with Python Django thanks to: Dennis Ivy from Python Django 7 Hour Course on YT

+++ studybud1/base/views.py
@@ -90,7 +90,7 @@ def updateRoom(request, pk):
     room = Room.objects.get(id=pk)
     form = RoomForm(instance=room)
 
-    if request.user != room.host:
+    if request.user != room.host and request.user != room.host:
         return HttpResponse("Your are not allowed here!!")
 
     if request.method == "POST":
@@ -106,6 +106,9 @@ def updateRoom(request, pk):
 @login_required(login_url="/login")
 def deleteRoom(request, pk):
     room = Room.objects.get(id=pk)
+
+    if request.user != room.host and request.user != room.host:
+        return HttpResponse("Your are not allowed here!!")
 
     if request.method == "POST":
         room.delete()
